% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overplot_data.R
\name{overplot_data}
\alias{overplot_data}
\title{Random Data Generator}
\usage{
overplot_data(n = 10000)
}
\arguments{
\item{n}{Number of rows. Default 10000}
}
\value{
Dataframe
}
\description{
A simple function for generating a data frame with two columns with
\code{n} rows.
}
