% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raster_example}
\alias{raster_example}
\title{Raster example}
\usage{
data(raster_example)
}
\value{
A matrix
}
\description{
Simple matrix used to illustrate geom_raster
}
\examples{
set.seed(1)
raster_example = expand.grid(x=1:10, y=1:10)
raster_example$z = runif(100)
}
\keyword{datasets}
